+++ spec/lib/http_server_manager/server_spec.rb
@@ -14,19 +14,19 @@ describe HttpServerManager::Server do
 
   describe "#start!" do
 
-    before(:each) { allow(Wait).to receive(:until_true!) }
+    before(:example) { allow(Wait).to receive(:until_true!) }
 
     context "when the server is not running" do
 
       let(:env_timeout) { nil }
 
-      before(:each) do
+      before(:example) do
         allow(ENV).to receive(:[]).with("timeout").and_return(env_timeout)
         allow(server).to receive(:running?).and_return(false)
         allow(Process).to receive(:spawn)
       end
 
-      it "should start the server by spawning a process that executes the start command" do
+      it "starts the server by spawning a process that executes the start command" do
         expect(Process).to receive(:spawn).with("some command", anything).and_return(888)
 
         server.start!
@@ -36,7 +36,7 @@ describe HttpServerManager::Server do
 
         let(:server_options) { { name: "Test Server", host: "localhost", port: 8888, timeout_in_seconds: 3 } }
 
-        it "should wait for the specified amount of time in seconds for the server to start" do
+        it "waits for the specified amount of time in seconds for the server to start" do
           expect(Wait).to receive(:until_true!).with(hash_including(timeout_in_seconds: 3))
 
           server.start!
@@ -48,7 +48,7 @@ describe HttpServerManager::Server do
 
         let(:env_timeout) { "8" }
 
-        it "should wait for the specified amount of time in seconds for the server to start" do
+        it "waits for the specified amount of time in seconds for the server to start" do
           expect(Wait).to receive(:until_true!).with(hash_including(timeout_in_seconds: 8))
 
           server.start!
@@ -58,7 +58,7 @@ describe HttpServerManager::Server do
 
       context "when no timeout is provided" do
 
-        it "should wait 20 seconds for the server to start" do
+        it "waits 20 seconds for the server to start" do
           expect(Wait).to receive(:until_true!).with(hash_including(timeout_in_seconds: 20))
 
           server.start!
@@ -72,12 +72,12 @@ describe HttpServerManager::Server do
 
   describe "#restart!" do
 
-    before(:each) do
+    before(:example) do
       allow(server).to receive(:start!)
       allow(server).to receive(:stop!)
     end
 
-    it "should first stop the server and then start the server" do
+    it "stops the server first and then starts the server" do
       expect(server).to receive(:stop!).ordered
       expect(server).to receive(:start!).ordered